+++ lib/src/_scaffolding/error/improper_user.dart
@@ -1,0 +1,15 @@
+library blackboard.rest.error.user;
+
+import 'rest_exception.dart';
+
+/// The [ImproperUser] class...
+class ImproperUser extends BlackboardRestException {
+  /// The [ImproperUser] constructor...
+  ImproperUser (
+    String message,
+    {String status, String code, String developerMessage, String extraInfo}
+  ) : super (
+    message, status: status, code: code,
+    developerMessage: developerMessage, extraInfo: extraInfo
+  );
+}