+++ lib/src/course_grades/bb_rest_course_grades.dart
@@ -5,8 +5,12 @@ import 'dart:async' show Future;
 import '../../course_grades.dart';
 
 import '../_scaffolding/configuration/endpoints.dart' show course_grades;
+
+import '../_scaffolding/error/invalid_grade.dart';
+
 import '../_scaffolding/factory/course_grades/column_attempt_factory.dart';
 import '../_scaffolding/factory/course_grades/grade_column_factory.dart';
+
 import '../_scaffolding/bb_rest_services.dart';
 
 import '../oauth2/access_token.dart';
@@ -21,17 +25,25 @@ class BbRestCourseGrades extends BlackboardRestServices implements CourseGrades 
   BbRestCourseGrades (String host, AccessToken token) : super (host, token);
 
   /// The [getGradeColumns] method...
+  @override
   Future<Iterable<GradeColumn>> getGradeColumns (String courseId) async {
     String endpoint = course_grades['get_grade_columns']
       .replaceFirst ('{courseId}', courseId);
 
-    Iterable<Object> rawResults =
-      (await connector.sendBbRestRequest (endpoint, useVersion: 2) as Map)['results'];
+    Iterable<Object> rawResults;
+
+    try {
+      rawResults =
+        (await connector.sendBbRestRequest (endpoint, useVersion: 2) as Map)['results'];
+    } catch (e) {
+      throw e as InvalidGrade;
+    }
 
     return (new GradeColumnFactory()).createAll (rawResults.cast());
   }
 
   /// The [createGradeColumn] method...
+  @override
   Future<void> createGradeColumn (String courseId, GradeColumn gradeColumn) async {}
 
   /// The [getGradeColumn] method...
@@ -40,17 +52,25 @@ class BbRestCourseGrades extends BlackboardRestServices implements CourseGrades 
       .replaceFirst ('{courseId}', courseId)
       .replaceFirst ('{columnId}', columnId);
 
-    Object rawResult = await connector.sendBbRestRequest (endpoint, useVersion: 2);
+    Object rawResult;
+
+    try {
+      rawResult = await connector.sendBbRestRequest (endpoint, useVersion: 2);
+    } catch (e) {
+      throw e as InvalidGrade;
+    }
 
     return (new GradeColumnFactory()).create (rawResult);
   }
 
   /// The [updateGradeColumn] method...
+  @override
   Future<void> updateGradeColumn (
     String courseId, String columnId, GradeColumn gradeColumn
   ) async {}
 
   /// The [getColumnAttempts] method...
+  @override
   Future<Iterable<Attempt>> getColumnAttempts (
     String courseId, String columnId
   ) async {
@@ -58,16 +78,24 @@ class BbRestCourseGrades extends BlackboardRestServices implements CourseGrades 
       .replaceFirst ('{courseId}', courseId)
       .replaceFirst ('{columnId}', columnId);
 
-    Iterable<Object> rawResults =
-      (await connector.sendBbRestRequest (endpoint, useVersion: 2) as Map)['results'];
+    Iterable<Object> rawResults;
+
+    try {
+      rawResults =
+        (await connector.sendBbRestRequest (endpoint, useVersion: 2) as Map)['results'];
+    } catch (e) {
+      throw e as InvalidGrade;
+    }
 
     return (new ColumnAttemptFactory()).createAll (rawResults.cast());
   }
 
   /// The [createColumnAttempt] method...
+  @override
   Future<void> createColumnAttempt (String columnId, Attempt attempt) async {}
 
   /// The [getColumnAttempt] method...
+  @override
   Future<Attempt> getColumnAttempt (
     String courseId, String columnId, String attemptId
   ) async {
@@ -76,30 +104,42 @@ class BbRestCourseGrades extends BlackboardRestServices implements CourseGrades 
       .replaceFirst ('{columnId}', columnId)
       .replaceFirst ('{attemptId}', attemptId);
 
-    Object rawResult = await connector.sendBbRestRequest (endpoint, useVersion: 2);
+    Object rawResult;
+
+    try {
+      rawResult = await connector.sendBbRestRequest (endpoint, useVersion: 2);
+    } catch (e) {
+      throw e as InvalidGrade;
+    }
 
     return (new ColumnAttemptFactory()).create (rawResult);
   }
 
   /// The [updateColumnAttempt] method...
+  @override
   Future<void> updateColumnAttempt (
     String columnId, String attemptId, Attempt attempt
   ) async {}
 
   /// The [getColumnGrades] method...
+  @override
   Future<Iterable<Grade>> getColumnGrades (String courseId, String columnId) async {}
 
   /// The [getColumnGradeLastChanged] method...
+  @override
   Future<Grade> getColumnGradeLastChanged (String courseId, String columnId) async {}
 
   /// The [getColumnGrade] method...
+  @override
   Future<Grade> getColumnGrade (String courseId, String columnId, String userId) async {}
 
   /// The [updateColumnGrade] method...
+  @override
   Future<void> updateColumnGrade (
     String courseId, String columnId, String userId, Grade grade
   ) async {}
 
   /// The [getUserGrades] method...
+  @override
   Future<Iterable<Grade>> getUserGrades (String courseId, String userId) async {}
 }