+++ lib/src/users/bb_rest_users.dart
@@ -1,0 +1,60 @@
+library blackboard.rest.implementation.users;
+
+import 'dart:async' show Future;
+
+import '../../users.dart';
+
+import '../_scaffolding/configuration/endpoints.dart' show users;
+import '../_scaffolding/error/improper_user.dart';
+import '../_scaffolding/factory/user_factory.dart';
+import '../_scaffolding/bb_rest_services.dart';
+
+import '../oauth2/access_token.dart';
+
+import 'user.dart';
+
+/// The [BbRestUsers] class...
+class BbRestUsers extends BlackboardRestServices implements Users {
+  /// The [BbRestUsers] constructor...
+  BbRestUsers (String host, AccessToken token) : super (host, token);
+
+  /// The [getUsers] method...
+  @override
+  Future<Iterable<User>> getUsers() async {
+    String endpoint = users['users'];
+
+    Iterable<Object> rawResults;
+
+    try {
+      rawResults = (await connector.sendBbRestRequest (endpoint) as Map)['results'];
+    } catch (e) {
+      throw e as ImproperUser;
+    }
+
+    return (new UserFactory()).createAll (rawResults.cast());
+  }
+
+  /// The [createUser] method...
+  @override
+  Future<void> createUser (User user) async {}
+
+  /// The [getUser] method...
+  @override
+  Future<User> getUser (String userId) async {
+    String endpoint = users['user'].replaceFirst ('{userId}', userId);
+
+    Object rawResult;
+
+    try {
+      rawResult = await connector.sendBbRestRequest (endpoint);
+    } catch (e) {
+      throw e as ImproperUser;
+    }
+
+    return (new UserFactory()).create (rawResult);
+  }
+
+  /// The [updateUser] method...
+  @override
+  Future<void> updateUser (String userId, User user) async {}
+}