+++ lib/users.dart
@@ -2,7 +2,14 @@ library blackboard.rest.users;
 
 import 'dart:async' show Future;
 
+import 'src/oauth2/access_token.dart';
+
+import 'src/users/bb_rest_users.dart';
 import 'src/users/user.dart';
+
+/// The [getUsersInstance] function...
+Users getUsersInstance (String host, AccessToken token) =>
+  new BbRestUsers (host, token);
 
 /// The [Users] abstract class...
 abstract class Users {